requirements.txt
streamlit>=1.32
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai>=0.1.5
langchain-text-splitters==0.2.4
faiss-cpu>=1.7.4
pypdf>=4.2
python-dotenv>=1.0
sentence-transformers>=2.7.0
